pylint: fix pylint 2.14 issues
Fix various issues detected by pylint 2.14 in the github action:

Convert open() to io.open() with an encoding,
and use  with  instead of try .. finally:
  W1514: Using open without explicitly specifying an encoding (unspecified-encoding)
  R1732: Consider using 'with' for resource-allocating operations (consider-using-with)

Convert old-style format strings to "{0}".format().
(Use "{0}" instead of "{}" for python 2.6 compatibility):
  C0209: Formatting a regular string which could be a f-string (consider-using-f-string)

Disable R1732 (consider-using-with) for subprocess.Popen() because
that doesn't support the context manager in python < 3.2.